project/backend/requirements.txt
Flask==3.0.0
mysql-connector-python==8.0.28
Werkzeug==2.2.2


